threat-feed-aggregator/requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Werkzeug==3.1.4
requests==2.32.4
aiohttp==3.13.2
APScheduler==3.10.4
SQLAlchemy==2.0.45
ldap3==2.9.1
geoip2==4.7.0
python-whois==0.9.6
tzlocal==5.2
gunicorn==23.0.0
cryptography==44.0.1
dnspython==2.6.1
aiodns==3.1.1
pycares==4.9.0
pytz
pyotp
qrcode
Pillow
aiohttp
aiodns
charset-normalizer
Flask
Flask-Login
Flask-Session
requests
schedule
beautifulsoup4
dnspython
ldap3
pyOpenSSL
python-whois
openpyxl
pandas
maxminddb
geoip2
apscheduler
SQLAlchemy
gunicorn
uvloop
psycopg2-binary
redis
pytest